pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "build"]
build-backend = "setuptools.build_meta"

[project]
name = "langextract-bedrock"
version = "0.1.0"
description = "AWS Bedrock provider plugin for LangExtract"
readme = "README.md"
license = { text = "Apache-2.0" }
authors = [{ name = "Vaibhavi Jhawar" }]
keywords = ["langextract", "aws", "bedrock", "llm"]
requires-python = ">=3.9"
dependencies = [
  "langextract>=1.0.0",
  "boto3>=1.34.0",
  "botocore>=1.34.0",
  "python-dotenv>=1.0.0"
]

[project.urls]
Homepage = "https://github.com/vjhawar-sp/langextract-bedrock"
Issues = "https://github.com/vjhawar-sp/langextract-bedrock/issues"

[project.entry-points."langextract.providers"]
bedrock = "langextract_bedrock:BedrockLanguageModel"
